+++ src/rustup.ts
@@ -81,10 +81,20 @@ export async function checkForRls(config: RustupConfig) {
   }
 }
 
-async function hasToolchain(config: RustupConfig): Promise<boolean> {
+async function hasToolchain({ channel, path }: RustupConfig): Promise<boolean> {
+  // In addition to a regular channel name, also handle shorthands e.g.
+  // `stable-msvc` or `stable-x86_64-msvc` but not `stable-x86_64-pc-msvc`.
+  const abiSuffix = ['-gnu', '-msvc'].find(abi => channel.endsWith(abi));
+  const [prefix, suffix] =
+    abiSuffix && channel.split('-').length <= 3
+      ? [channel.substr(0, channel.length - abiSuffix.length), abiSuffix]
+      : [channel, undefined];
+  // Skip middle target triple components such as vendor as necessary, since
+  // `rustup` output lists toolchains with a full target triple inside
+  const matcher = new RegExp([prefix, suffix && `.*${suffix}`].join(''));
   try {
-    const { stdout } = await exec(`${config.path} toolchain list`);
-    return stdout.includes(config.channel);
+    const { stdout } = await exec(`${path} toolchain list`);
+    return matcher.test(stdout);
   } catch (e) {
     console.log(e);
     window.showErrorMessage(