+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### Unreleased
 
+* Accept rustup toolchain shorthands in `rust-client.channel`, e.g. `stable-gnu` or `nightly-x86_64-msvc`
 * Remove deprecated `rust-client.useWsl` setting (use the official
 [Remote - WSL](https://marketplace.visualstudio.com/items?itemName=ms-vscode-remote.remote-wsl) extension instead)